feat:일반 응답 객체 및 응답 코드, 방 응답 코드 및 객체, DTO 생성 #BACK-124 Related to : #124

+++ backend/src/main/java/signiel/heartsigniel/model/party/Party.java
@@ -25,7 +25,7 @@ public class Party {
     private String partyGender;
 
     @Column(name="avg_rating")
-    private int avgRating;
+    private Long avgRating;
 
     @Column(name = "party_type", length = 20)
     private String partyType;
